cubeprop cubed to cubeprop solo
