+++ src/bin/psi4/CMakeLists.txt
@@ -30,7 +30,7 @@ list(SORT headers_list)
 
 set(sources_list "")
 # List of sources
-list(APPEND sources_list export_psio.cc export_mints.cc psi_stop.cc export_functional.cc export_oeprop.cc export_plugins.cc export_blas_lapack.cc export_cubefile.cc export_benchmarks.cc export_efp.cc export_cubeprop.cc clean.cc create_new_plugin.cc script.cc set_memory.cc read_options.cc export_libparallel.cc)
+list(APPEND sources_list export_psio.cc export_mints.cc psi_stop.cc export_functional.cc export_oeprop.cc export_plugins.cc export_blas_lapack.cc export_benchmarks.cc export_efp.cc export_cubeprop.cc clean.cc create_new_plugin.cc script.cc set_memory.cc read_options.cc export_libparallel.cc)
 
 # If you want to remove some sources specify them explictly here
 if(DEVELOPMENT_CODE)
